Add benchmark.
Upon viewing the benchmark results (see src/main.cpp), the vector template specialization of the SDR
outperforms the list and set specializations by a couple orders of magnitude (focusing on ands, since it will be used a lot),
even if the size of the SDR is very large (~6000 elements). Although
the performance may be increase by ~30% if the approperiate boost
intrusive set and slist containers are used instead, I just don't see it worth the time
and effort to pursue the underperforming versions further. Considering the removal
of the container template param from SDR, and leaving only the vector version.